telegram_bot/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
PyJWT
cryptography
sentence-transformers
python-dotenv
python-telegram-bot
slowapi
python-telegram-bot[rate-limiter]
